Include MAC by default if encryption supports it
DEVSIX-8622

Autoported commit.
Original commit hash: [209419420]
Manual files:
sharpenConfiguration.xml

+++ itext/itext.kernel/itext/kernel/crypto/securityhandler/StandardHandlerUsingStandard40.cs
@@ -163,7 +163,7 @@ namespace iText.Kernel.Crypto.Securityhandler {
             byte[] uValue = GetIsoBytes(encryptionDictionary.GetAsString(PdfName.U));
             byte[] oValue = GetIsoBytes(encryptionDictionary.GetAsString(PdfName.O));
             PdfNumber pValue = (PdfNumber)encryptionDictionary.Get(PdfName.P);
-            this.permissions = pValue.LongValue();
+            this.permissions = pValue.IntValue();
             this.documentId = documentId;
             keyLength = GetKeyLength(encryptionDictionary);
             byte[] paddedPassword = PadPassword(password);
